requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.1
django-braces==1.12.0
django-crispy-forms==1.7.0
django-model-utils==3.1.1
django-rest-framework==0.1.0
django-star-ratings==0.6.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
swapper==1.1.0
urllib3==1.22
